backend/pyproject.toml
[project]
name = "okrtracker"
version = "0.1.0"
description = "OKR tracker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers==4.9.0",
    "django==6.0.3",
    "django-environ==0.13.0",
    "django-split-settings==1.3.2",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "psycopg[binary]>=3.3.3",
]

[dependency-groups]
dev = [
    "django-stubs==5.2.9",
    "djangorestframework-stubs==3.16.8",
    "model-bakery==1.23.3",
    "pre-commit==4.5.1",
    "pytest==9.0.2",
    "pytest-django==4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["core/tests", "okr/tests"]
addopts = "-ra"

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = [
    "__pycache__",
    ".pytest_cache",
    ".ruff_cache",
    "**/generated/**",
    "**/migrations/**",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D106", "D107", "D212"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D205"]
"core/tests/**/*.py" = ["D205"]
"okr/tests/**/*.py" = ["D205"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
skip-magic-trailing-comma = false

[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
include = ["config", "manage.py"]
respect-ignore-files = true
exclude = [
    "__pycache__",
    ".pytest_cache",
    ".ruff_cache",
    "**/generated/**",
    "**/migrations/**",
]
